Revert "fix bugs with cookies"
This reverts commit a2a0bdc3fc891721a8340342e5a728cd1129d9e6.

Turns out my browser always recommended the URL with a . at the end of
the hostname, which resulted in different cookies being used. The
redirect would then happen to the normal URL, without the trailing .,
which had the already authenticated cookie saved. Due to the different
nonces for the different sessions, the result was an internal server
error.